requirements.txt
# 필요한 패키지들
streamlit
openai
python-dotenv